+++ libs/libc/machine/arch_atomic_irq.c
@@ -28,291 +28,24 @@
 
 #include <stdbool.h>
 #include <stdint.h>
-#include <nuttx/spinlock.h>
+#include <nuttx/irq.h>
 #include <nuttx/macro.h>
 
+#include "arch_atomic.h"
+
 /****************************************************************************
- * Pre-processor Definitions
+ * Private Functions
  ****************************************************************************/
 
-#define STORE(fn, n, type)                                           \
-                                                                     \
-  void weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-                                                                     \
-    *(FAR type *)ptr = value;                                        \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-  }
+static inline irqstate_t atomic_lock(void)
+{
+  return up_irq_save();
+}
 
-#define LOAD(fn, n, type)                                             \
-                                                                      \
-  type weak_function CONCATENATE(fn, n)(FAR const volatile void *ptr, \
-                                        int memorder)                 \
-  {                                                                   \
-    irqstate_t irqstate = up_irq_save();                              \
-                                                                      \
-    type ret = *(FAR type *)ptr;                                      \
-                                                                      \
-    up_irq_restore(irqstate);                                         \
-    return ret;                                                       \
-  }
-
-#define EXCHANGE(fn, n, type)                                        \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    type ret = *tmp;                                                 \
-    *tmp = value;                                                    \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define CMP_EXCHANGE(fn, n, type)                                    \
-                                                                     \
-  bool weak_function CONCATENATE(fn, n)(FAR volatile void *mem,      \
-                                        FAR volatile void *expect,   \
-                                        type desired, bool weak,     \
-                                        int success, int failure)    \
-  {                                                                  \
-    bool ret = false;                                                \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmpmem = (FAR type *)mem;                              \
-    FAR type *tmpexp = (FAR type *)expect;                           \
-                                                                     \
-    if (*tmpmem == *tmpexp)                                          \
-      {                                                              \
-        ret = true;                                                  \
-        *tmpmem = desired;                                           \
-      }                                                              \
-    else                                                             \
-      {                                                              \
-        *tmpexp = *tmpmem;                                           \
-      }                                                              \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FLAG_TEST_AND_SET(fn, n, type)                               \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        int memorder)                \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *(FAR type *)ptr = 1;                                            \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FETCH_ADD(fn, n, type)                                       \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *tmp = *tmp + value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FETCH_SUB(fn, n, type)                                       \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *tmp = *tmp - value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FETCH_AND(fn, n, type)                                       \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *tmp = *tmp & value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FETCH_OR(fn, n, type)                                        \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *tmp = *tmp | value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define FETCH_XOR(fn, n, type)                                       \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value, int memorder)    \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    *tmp = *tmp ^ value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define SYNC_ADD_FETCH(fn, n, type)                                  \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = *tmp + value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_SUB_FETCH(fn, n, type)                                  \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = *tmp - value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_OR_FETCH(fn, n, type)                                   \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = *tmp | value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_AND_FETCH(fn, n, type)                                  \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = *tmp & value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_XOR_FETCH(fn, n, type)                                  \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = *tmp ^ value;                                             \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_NAND_FETCH(fn, n, type)                                 \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type value)                  \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    *tmp = ~(*tmp & value);                                          \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return *tmp;                                                     \
-  }
-
-#define SYNC_BOOL_CMP_SWAP(fn, n, type)                              \
-                                                                     \
-  bool weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type oldvalue,               \
-                                        type newvalue)               \
-  {                                                                  \
-    bool ret = false;                                                \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-                                                                     \
-    if (*tmp == oldvalue)                                            \
-      {                                                              \
-        ret = true;                                                  \
-        *tmp = newvalue;                                             \
-      }                                                              \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
-
-#define SYNC_VAL_CMP_SWAP(fn, n, type)                               \
-                                                                     \
-  type weak_function CONCATENATE(fn, n)(FAR volatile void *ptr,      \
-                                        type oldvalue,               \
-                                        type newvalue)               \
-  {                                                                  \
-    irqstate_t irqstate = up_irq_save();                             \
-    FAR type *tmp = (FAR type *)ptr;                                 \
-    type ret = *tmp;                                                 \
-                                                                     \
-    if (*tmp == oldvalue)                                            \
-      {                                                              \
-        *tmp = newvalue;                                             \
-      }                                                              \
-                                                                     \
-    up_irq_restore(irqstate);                                        \
-    return ret;                                                      \
-  }
+static inline void atomic_unlock(irqstate_t flags)
+{
+  up_irq_restore(flags);
+}
 
 /****************************************************************************
  * Public Functions
@@ -338,37 +71,23 @@ STORE(__atomic_store_, 4, uint32_t)
 STORE(nx_atomic_store_, 4, int32_t)
 
 /****************************************************************************
- * Name: __atomic_store_8
- ****************************************************************************/
-
-STORE(__atomic_store_, 8, uint64_t)
-STORE(nx_atomic_store_, 8, int64_t)
-
-/****************************************************************************
  * Name: __atomic_load_1
  ****************************************************************************/
 
 LOAD(__atomic_load_, 1, uint8_t)
 
 /****************************************************************************
- * Name: __atomic_load__2
+ * Name: __atomic_load_2
  ****************************************************************************/
 
 LOAD(__atomic_load_, 2, uint16_t)
 
 /****************************************************************************
- * Name: __atomic_load__4
+ * Name: __atomic_load_4
  ****************************************************************************/
 
 LOAD(__atomic_load_, 4, uint32_t)
 LOAD(nx_atomic_load_, 4, int32_t)
-
-/****************************************************************************
- * Name: __atomic_load__8
- ****************************************************************************/
-
-LOAD(__atomic_load_, 8, uint64_t)
-LOAD(nx_atomic_load_, 8, int64_t)
 
 /****************************************************************************
  * Name: __atomic_exchange_1
@@ -377,24 +96,17 @@ LOAD(nx_atomic_load_, 8, int64_t)
 EXCHANGE(__atomic_exchange_, 1, uint8_t)
 
 /****************************************************************************
- * Name: __atomic_exchange__2
+ * Name: __atomic_exchange_2
  ****************************************************************************/
 
 EXCHANGE(__atomic_exchange_, 2, uint16_t)
 
 /****************************************************************************
- * Name: __atomic_exchange__4
+ * Name: __atomic_exchange_4
  ****************************************************************************/
 
 EXCHANGE(__atomic_exchange_, 4, uint32_t)
 EXCHANGE(nx_atomic_exchange_, 4, int32_t)
-
-/****************************************************************************
- * Name: __atomic_exchange__8
- ****************************************************************************/
-
-EXCHANGE(__atomic_exchange_, 8, uint64_t)
-EXCHANGE(nx_atomic_exchange_, 8, int64_t)
 
 /****************************************************************************
  * Name: __atomic_compare_exchange_1
@@ -416,13 +128,6 @@ CMP_EXCHANGE(__atomic_compare_exchange_, 4, uint32_t)
 CMP_EXCHANGE(nx_atomic_compare_exchange_, 4, int32_t)
 
 /****************************************************************************
- * Name: __atomic_compare_exchange_8
- ****************************************************************************/
-
-CMP_EXCHANGE(__atomic_compare_exchange_, 8, uint64_t)
-CMP_EXCHANGE(nx_atomic_compare_exchange_, 8, int64_t)
-
-/****************************************************************************
  * Name: __atomic_flag_test_and_set_1
  ****************************************************************************/
 
@@ -440,13 +145,6 @@ FLAG_TEST_AND_SET(__atomic_flags_test_and_set_, 2, uint16_t)
 
 FLAG_TEST_AND_SET(__atomic_flags_test_and_set_, 4, uint32_t)
 FLAG_TEST_AND_SET(nx_atomic_flags_test_and_set_, 4, int32_t)
-
-/****************************************************************************
- * Name: __atomic_flag_test_and_set_8
- ****************************************************************************/
-
-FLAG_TEST_AND_SET(__atomic_flags_test_and_set_, 8, uint64_t)
-FLAG_TEST_AND_SET(nx_atomic_flags_test_and_set_, 8, int64_t)
 
 /****************************************************************************
  * Name: __atomic_fetch_add_1
@@ -468,13 +166,6 @@ FETCH_ADD(__atomic_fetch_add_, 4, uint32_t)
 FETCH_ADD(nx_atomic_fetch_add_, 4, int32_t)
 
 /****************************************************************************
- * Name: __atomic_fetch_add_8
- ****************************************************************************/
-
-FETCH_ADD(__atomic_fetch_add_, 8, uint64_t)
-FETCH_ADD(nx_atomic_fetch_add_, 8, int64_t)
-
-/****************************************************************************
  * Name: __atomic_fetch_sub_1
  ****************************************************************************/
 
@@ -492,13 +183,6 @@ FETCH_SUB(__atomic_fetch_sub_, 2, uint16_t)
 
 FETCH_SUB(__atomic_fetch_sub_, 4, uint32_t)
 FETCH_SUB(nx_atomic_fetch_sub_, 4, int32_t)
-
-/****************************************************************************
- * Name: __atomic_fetch_sub_8
- ****************************************************************************/
-
-FETCH_SUB(__atomic_fetch_sub_, 8, uint64_t)
-FETCH_SUB(nx_atomic_fetch_sub_, 8, int64_t)
 
 /****************************************************************************
  * Name: __atomic_fetch_and_1
@@ -520,13 +204,6 @@ FETCH_AND(__atomic_fetch_and_, 4, uint32_t)
 FETCH_AND(nx_atomic_fetch_and_, 4, int32_t)
 
 /****************************************************************************
- * Name: __atomic_fetch_and_8
- ****************************************************************************/
-
-FETCH_AND(__atomic_fetch_and_, 8, uint64_t)
-FETCH_AND(nx_atomic_fetch_and_, 8, int64_t)
-
-/****************************************************************************
  * Name: __atomic_fetch_or_1
  ****************************************************************************/
 
@@ -546,13 +223,6 @@ FETCH_OR(__atomic_fetch_or_, 4, uint32_t)
 FETCH_OR(nx_atomic_fetch_or_, 4, int32_t)
 
 /****************************************************************************
- * Name: __atomic_fetch_or_4
- ****************************************************************************/
-
-FETCH_OR(__atomic_fetch_or_, 8, uint64_t)
-FETCH_OR(nx_atomic_fetch_or_, 8, int64_t)
-
-/****************************************************************************
  * Name: __atomic_fetch_xor_1
  ****************************************************************************/
 
@@ -570,13 +240,6 @@ FETCH_XOR(__atomic_fetch_xor_, 2, uint16_t)
 
 FETCH_XOR(__atomic_fetch_xor_, 4, uint32_t)
 FETCH_XOR(nx_atomic_fetch_xor_, 4, int32_t)
-
-/****************************************************************************
- * Name: __atomic_fetch_xor_8
- ****************************************************************************/
-
-FETCH_XOR(__atomic_fetch_xor_, 8, uint64_t)
-FETCH_XOR(nx_atomic_fetch_xor_, 8, int64_t)
 
 /* Clang define the __sync builtins, add #ifndef to avoid
  * redefined/redeclared problem.
@@ -603,12 +266,6 @@ SYNC_ADD_FETCH(__sync_add_and_fetch_, 2, uint16_t)
 SYNC_ADD_FETCH(__sync_add_and_fetch_, 4, uint32_t)
 
 /****************************************************************************
- * Name: __sync_add_and_fetch_8
- ****************************************************************************/
-
-SYNC_ADD_FETCH(__sync_add_and_fetch_, 8, uint64_t)
-
-/****************************************************************************
  * Name: __sync_sub_and_fetch_1
  ****************************************************************************/
 
@@ -625,12 +282,6 @@ SYNC_SUB_FETCH(__sync_sub_and_fetch_, 2, uint16_t)
  ****************************************************************************/
 
 SYNC_SUB_FETCH(__sync_sub_and_fetch_, 4, uint32_t)
-
-/****************************************************************************
- * Name: __sync_sub_and_fetch_8
- ****************************************************************************/
-
-SYNC_SUB_FETCH(__sync_sub_and_fetch_, 8, uint64_t)
 
 /****************************************************************************
  * Name: __sync_or_and_fetch_1
@@ -651,12 +302,6 @@ SYNC_OR_FETCH(__sync_or_and_fetch_, 2, uint16_t)
 SYNC_OR_FETCH(__sync_or_and_fetch_, 4, uint32_t)
 
 /****************************************************************************
- * Name: __sync_or_and_fetch_8
- ****************************************************************************/
-
-SYNC_OR_FETCH(__sync_or_and_fetch_, 8, uint64_t)
-
-/****************************************************************************
  * Name: __sync_and_and_fetch_1
  ****************************************************************************/
 
@@ -673,12 +318,6 @@ SYNC_AND_FETCH(__sync_and_and_fetch_, 2, uint16_t)
  ****************************************************************************/
 
 SYNC_AND_FETCH(__sync_and_and_fetch_, 4, uint32_t)
-
-/****************************************************************************
- * Name: __sync_and_and_fetch_8
- ****************************************************************************/
-
-SYNC_AND_FETCH(__sync_and_and_fetch_, 8, uint64_t)
 
 /****************************************************************************
  * Name: __sync_xor_and_fetch_1
@@ -699,12 +338,6 @@ SYNC_XOR_FETCH(__sync_xor_and_fetch_, 2, uint16_t)
 SYNC_XOR_FETCH(__sync_xor_and_fetch_, 4, uint32_t)
 
 /****************************************************************************
- * Name: __sync_xor_and_fetch_8
- ****************************************************************************/
-
-SYNC_XOR_FETCH(__sync_xor_and_fetch_, 8, uint64_t)
-
-/****************************************************************************
  * Name: __sync_nand_and_fetch_1
  ****************************************************************************/
 
@@ -721,12 +354,6 @@ SYNC_NAND_FETCH(__sync_nand_and_fetch_, 2, uint16_t)
  ****************************************************************************/
 
 SYNC_NAND_FETCH(__sync_nand_and_fetch_, 4, uint32_t)
-
-/****************************************************************************
- * Name: __sync_nand_and_fetch_8
- ****************************************************************************/
-
-SYNC_NAND_FETCH(__sync_nand_and_fetch_, 8, uint64_t)
 
 /****************************************************************************
  * Name: __sync_bool_compare_and_swap_1
@@ -747,12 +374,6 @@ SYNC_BOOL_CMP_SWAP(__sync_bool_compare_and_swap_, 2, uint16_t)
 SYNC_BOOL_CMP_SWAP(__sync_bool_compare_and_swap_, 4, uint32_t)
 
 /****************************************************************************
- * Name: __sync_bool_compare_and_swap_8
- ****************************************************************************/
-
-SYNC_BOOL_CMP_SWAP(__sync_bool_compare_and_swap_, 8, uint64_t)
-
-/****************************************************************************
  * Name: __sync_val_compare_and_swap_1
  ****************************************************************************/
 
@@ -769,12 +390,6 @@ SYNC_VAL_CMP_SWAP(__sync_val_compare_and_swap_, 2, uint16_t)
  ****************************************************************************/
 
 SYNC_VAL_CMP_SWAP(__sync_val_compare_and_swap_, 4, uint32_t)
-
-/****************************************************************************
- * Name: __sync_val_compare_and_swap_8
- ****************************************************************************/
-
-SYNC_VAL_CMP_SWAP(__sync_val_compare_and_swap_, 8, uint64_t)
 
 /****************************************************************************
  * Name: __sync_synchronize
